+++ source/extension/unity/Package/Editor/KesManagerEditor.cs
@@ -1,6 +1,7 @@
 using UnityEditor;
 
-namespace KoromoEventScript.Unity.Editor;
+namespace KoromoEventScript.Unity.Editor
+{
 
 [CustomEditor(typeof(KesManager))]
 internal sealed class KesManagerEditor : UnityEditor.Editor
@@ -8,8 +9,13 @@ internal sealed class KesManagerEditor : UnityEditor.Editor
     public override void OnInspectorGUI()
     {
         DrawDefaultInspector();
-        EditorGUILayout.HelpBox(
-            "KES Build Asset support will be added with the runtime importer implementation.",
-            MessageType.Info);
+        var manager = (KesManager)target;
+        if (manager.BuildAsset == null)
+        {
+            EditorGUILayout.HelpBox(
+                "Assign a KES Build Asset imported from manifest.kson.",
+                MessageType.Warning);
+        }
     }
+}
 }